requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
langgraph==0.2.38
langchain==0.3.4
langchain-groq==0.2.1
langchain-core==0.3.12
pinecone-client==5.0.1
redis==5.1.1
aioredis==2.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
asyncio-redis==0.16.0
pytest==8.3.3
pytest-asyncio==0.24.0
aiofiles==24.1.0
jinja2==3.1.4
python-multipart==0.0.12
